+++ src/main/java/com/shopspace/shopspaceservice/service/ProductService.java
@@ -19,5 +19,5 @@ public interface ProductService {
     // Todo: methods for web site
     List<Product> getLastProductsWithLimit(Integer page, Integer size);
 
-    List<Product> getProductsByCategoryWithLimit(Integer category, Integer page, Integer size);
+    List<Product> getProductsByCategoryWithLimit(Long category, Integer page, Integer size);
 }
